Dev: ctags / cscope added as a build target
This build target requires exuberant ctags and cscope in order
to work. Generated files are out of source tree.

The local vim config is altered to find tags and cscope.out. It
is assumed, that the build directoy is right underneath the projects
root directory. The reason for this is simple, it is most convinient
for me, but it does not work if the build directory is located
elsewhere.