Math: initial import of the math package
This is yet another math package, containing vectors, matrices and
the quaternion. Certenly, there are other libraries out there, and
far better than this implementation. However, because this library
shall keep the number of dependencies low, this implementation is
necessary.

It does not provide as many functions as other implementations, but
this is enough for its purpose.