AIS: removed duplicate specification of offset and size of data within payload.
The mapping of data within the AIS message payload is done (almost completly)
at one point. It resolves at compile time and works for both, `get` and `set`
operations from and to the bitset that represents the AIS message.

This enhances the robustness of the code and it also gives a better overview
of the mapping.