NMEA: changed interface of sentences containing position information.
The interface does not longer return raw data for lat/lon and their
respective hemisphere. Access to the position information is now
provided with the functions

 - get_latitude
 - get_longitude

which return geo::latitude and geo::longitude. This interface is
easier to use and more intuitive, as well as it encapsulates the
position encoding in the NMEA sentence.