AIS: latitude/longitude of messages are now handled using utils::optional.
This is necessary because the geo::latitude and geo::longitude objects
do not have a 'not defined' status, but this can happen in an AIS message.

This also solves the situation, where an 'not defined' lat/lon is
present in an AIS message, bug geo::Latitude/geo::Longitude cannot
handle the out of range value. An exception was thrown.