Doc: build status (Travis-CI) removed from README.
This is necessary because Travis-CI breaks randomly build.

Regular occasions are not being able to download Clang or
another package. There are also ocurrences of misbehaviour
of CMake for finding packages which are clearly mulfunctions
of Travis-CI, example:

  find_package(Qt5Core)

does not lead to CMake aborting with an error if not found.

At this point Travis-CI cannot be taken seriously for C++
projects and is considered nice-to-have. A test-script
for desired combinations of options and compilers will be
provided in the scope of the project at a future point
in time.