NMEA: naive implementation of locale independent string/double conversions
This implementation is rather naive/basic, non-perfomant implemenation of
a locale independent string/double conversions (double to string, string
to double).

strtod_l is used, if available.

Unit tests are performed to check for other locales than classic. The
tests are disabled, if the locale is not available.

Better implemenation should be considered using a different algorithm like
Grisu2, Ryu, Dragon4, etc.

This commit tackles issue #27