Build: add VC++ support
Adds basic and incomplete support for VC++ builds. Tests and tools
are available, however conan packaging is not tested, examples are not
supported and everying with marnav-io as well. Older versions of MSVC
are not tested either.

CMakePresets are in place to build and it easily:

    cmake --preset vs2022-release-shared
    cmake --build --preset vs2022-release-shared
    ctest --preset vs2022-release-shared

Requirements:
- VC++ build tools for 14.34 (_MSC_VER=1934, Visual Studio 2022)
  Toolkits for older version are not tested.
- Win10 SDK

CI build for windows-2022 has been added to GitHub Actions.