Build: add ARM and aarch64 target architectures
This patch adds the CMake presets for ARM and aarch64 target
architecture for building and testing.
If `binfmt_misc` is enabled on Linux hosts, the execution of unit tests
can be done by just using `ctest`. Otherwise, use `qemu-arm` or
`qemu-aarch64` respectively.

Currently (officially) supported for those target architectures are the
library and unit tests. Other parts may or may not work.